+++ src/natcap/invest/urban_nature_access/urban_nature_access.py
@@ -34,10 +34,28 @@ RADIUS_OPT_URBAN_NATURE = 'radius per urban nature class'
 RADIUS_OPT_POP_GROUP = 'radius per population group'
 POP_FIELD_REGEX = '^pop_'
 ID_FIELDNAME = 'adm_unit_id'
+
+_model_description = gettext(
+    """
+    The InVEST Urban Nature Access model allows for examination of access to
+    the recreational, social, psychological, and physical health benefits of
+    nature. Assessing nature-based recreation and its health benefits first
+    requires an understanding of both the "supply" of urban nature and
+    "demand" for urban nature. The supply is determined by the type, size,
+    proximity, and quality of urban nature, while the demand depends on
+    people's preferences or local policy requirements/urban planning
+    standards. The InVEST Urban Nature Access model provides a measure of
+    both supply and demand for urban nature, as well as the supply-demand
+    balance, which quantifies the extent to which supply meets demand at
+    the individual, administrative, and city levels.
+    """)
+
 MODEL_SPEC = spec.ModelSpec(
     model_id="urban_nature_access",
     model_title=gettext("Urban Nature Access"),
     userguide="urban_nature_access.html",
+    reporter="natcap.invest.urban_nature_access.reporter",
+    about=_model_description,
     validate_spatial_overlap=True,
     different_projections_ok=True,
     aliases=("una",),
@@ -426,7 +444,8 @@ MODEL_SPEC = spec.ModelSpec(
                 "The area of greenspace available within the defined radius,"
                 " weighted by the selected decay function."
             ),
-            created_if="search_radius_mode == 'radius per urban nature class'",
+            created_if=(
+                f"search_radius_mode == '{RADIUS_OPT_UNIFORM}'"),
             data_type=float,
             units=u.meter**2
         ),
@@ -897,9 +916,9 @@ def execute(args):
         split_population_fields = list(
             filter(lambda x: re.match(POP_FIELD_REGEX, x),
                    validation.load_fields_from_vector(
-                       args['admin_boundaries_vector_path'])))
+                       file_registry['reprojected_admin_boundaries'])))
 
-        if _geometries_overlap(args['admin_boundaries_vector_path']):
+        if _geometries_overlap(file_registry['reprojected_admin_boundaries']):
             LOGGER.warning(
                 "Some administrative boundaries overlap, which will affect "
                 "the accuracy of supply rasters per population group. ")