+++ tests/reports/test_sdr_ndr_template.py
@@ -25,6 +25,8 @@ def _get_render_args(model_spec):
     outputs_caption = ['results.tif:Results map.']
     intermediate_outputs_caption = ['stream.tif:Stream network map.']
     raster_group_caption = 'This is another test!'
+    lulc_legend_html = '<table class="test__lulc-legend-table"></table>'
+    lulc_caption = ['lulc.tif: LULC map']
 
     return {
         'report_script': model_spec.reporter,
@@ -36,6 +38,9 @@ def _get_render_args(model_spec):
         'model_description': model_spec.about,
         'timestamp': timestamp,
         'args_dict': args_dict,
+        'lulc_img_src': img_src,
+        'lulc_legend_html': lulc_legend_html,
+        'lulc_caption': lulc_caption,
         'inputs_img_src': img_src,
         'inputs_caption': inputs_caption,
         'outputs_img_src': img_src,
@@ -66,7 +71,7 @@ class SDR_NDR_TemplateTests(unittest.TestCase):
         soup = BeautifulSoup(html, BSOUP_HTML_PARSER)
 
         sections = soup.find_all(class_='accordion-section')
-        self.assertEqual(len(sections), 8)
+        self.assertEqual(len(sections), 9)
         self.assertEqual(
             soup.title.string, f'InVEST Results: {MODEL_SPEC.model_title}')
 
@@ -80,41 +85,19 @@ class SDR_NDR_TemplateTests(unittest.TestCase):
         soup = BeautifulSoup(html, BSOUP_HTML_PARSER)
 
         sections = soup.find_all(class_='accordion-section')
-        self.assertEqual(len(sections), 8)
+        self.assertEqual(len(sections), 9)
         self.assertEqual(
             soup.title.string, f'InVEST Results: {MODEL_SPEC.model_title}')
 
     def test_watershed_results_totals(self):
         """Totals should be rendered when passed to the render function."""
+        from natcap.invest.ndr.ndr import MODEL_SPEC
 
-        ws_vector_table = '<table class="test__results-table"></table>'
-        ws_vector_totals_table = '<table class="test__totals-table"></table>'
+        render_args = _get_render_args(MODEL_SPEC)
+        render_args['ws_vector_table'] = '<table class="test__results-table"></table>'
+        render_args['ws_vector_totals_table'] = '<table class="test__totals-table"></table>'
 
-        html = TEMPLATE.render(
-            report_script='natcap.invest.test.reporter',
-            invest_version='987.65.0',
-            report_filepath='sdr_ndr_report_test.html',
-            model_id='',
-            model_name='',
-            userguide_page='',
-            model_description='',
-            timestamp='',
-            args_dict={},
-            inputs_img_src='',
-            inputs_caption=[],
-            outputs_img_src='',
-            outputs_caption=[],
-            intermediate_outputs_heading='',
-            intermediate_outputs_img_src='',
-            intermediate_outputs_caption=[],
-            raster_group_caption='',
-            ws_vector_table=ws_vector_table,
-            ws_vector_totals_table=ws_vector_totals_table,
-            output_raster_stats_table='',
-            input_raster_stats_table='',
-            stats_table_note='',
-            model_spec_outputs=[],
-        )
+        html = TEMPLATE.render(**render_args)
 
         soup = BeautifulSoup(html, BSOUP_HTML_PARSER)
 
@@ -126,34 +109,13 @@ class SDR_NDR_TemplateTests(unittest.TestCase):
     def test_watershed_results_without_totals(self):
         """Totals should be not be rendered when there are none to render."""
 
-        ws_vector_table = '<table class="test__results-table"></table>'
-        ws_vector_totals_table = None
+        from natcap.invest.ndr.ndr import MODEL_SPEC
 
-        html = TEMPLATE.render(
-            report_script='natcap.invest.test.reporter',
-            invest_version='987.65.0',
-            report_filepath='sdr_ndr_report_test.html',
-            model_id='',
-            model_name='',
-            userguide_page='',
-            model_description='',
-            timestamp='',
-            args_dict={},
-            inputs_img_src='',
-            inputs_caption=[],
-            outputs_img_src='',
-            outputs_caption=[],
-            intermediate_outputs_heading='',
-            intermediate_outputs_img_src='',
-            intermediate_outputs_caption=[],
-            raster_group_caption='',
-            ws_vector_table=ws_vector_table,
-            ws_vector_totals_table=ws_vector_totals_table,
-            output_raster_stats_table='',
-            input_raster_stats_table='',
-            stats_table_note='',
-            model_spec_outputs=[],
-        )
+        render_args = _get_render_args(MODEL_SPEC)
+        render_args['ws_vector_table'] = '<table class="test__results-table"></table>'
+        render_args['ws_vector_totals_table'] = None
+
+        html = TEMPLATE.render(**render_args)
 
         soup = BeautifulSoup(html, BSOUP_HTML_PARSER)